LTF_DOCKER/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
paramiko==3.3.1
pysmb==1.2.9.1
webdavclient3==3.14.6
requests==2.26.0